re-instate 0.6.0 arduino constants.  moves the new 0.6.1 arduino defines from spark_wiring_arduino.h to Ardino.h.
